requirements.txt
click==7.1.2
flake8==3.8.3
Flask==1.1.2
Flask-Cors==3.0.8
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
Pillow==7.2.0
pycodestyle==2.6.0
pyflakes==2.2.0
six==1.15.0
Werkzeug==1.0.1
zipp==3.1.0
